refactor: move browser, server and utils code generation options into a single file.
This helps with code and logic readability.
