+++ packages/angular_devkit/build_angular/src/tools/esbuild/angular/compilation/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+export { AngularCompilation } from './angular-compilation';
+export { AotCompilation } from './aot-compilation';
+export { JitCompilation } from './jit-compilation';
+export { NoopCompilation } from './noop-compilation';
