+++ packages/angular_devkit/build_angular/src/tools/esbuild/compiler-plugin-options.ts
